+++ py3dcore/methods/__init__.py
@@ -1,4 +1,3 @@
 # -*- coding: utf-8 -*-
 
 from .abc_smc import ABC_SMC
-from .base import BaseFitter, generate_ensemble
