+++ py3dcore/models/toroidal/thin_torus.py
@@ -182,9 +182,9 @@ def thin_torus_gh(
 
         # ellipse circumference
         h = (delta - 1) ** 2 / (1 + delta) ** 2
-        E = np.pi * (1 + delta) * (1 + 3 * h / (10 + np.sqrt(4 - 3 * h)))
+        Efac = np.pi * (1 + delta) * (1 + 3 * h / (10 + np.sqrt(4 - 3 * h)))
 
-        t = Tfac * rho_1 / rho_0 / 2 / np.pi * np.sin(q1 / 2) ** 2
+        t = Tfac * rho_1 / rho_0 / 2 / np.pi * np.sin(q1 / 2) ** 2 / Efac
 
         denom = 1 + t**2 * q0**2
         bpsi = b_t / denom * fluxfactor
